+++ previous-versions/sync-for-expenses-version-1/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [go v0.1.0] previous-versions/sync-for-expenses-version-1
 ### Releases
 - [Go v0.1.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-expenses-version-1/v0.1.0 - previous-versions/sync-for-expenses-version-1
+
+## 2023-09-09 00:04:40
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses-v1.yaml
+- Speakeasy CLI 1.82.3 (2.107.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.24.0] previous-versions/sync-for-expenses-version-1
+### Releases
+- [Go v0.24.0] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/sync-for-expenses-version-1/v0.24.0 - previous-versions/sync-for-expenses-version-1